flight_pattern_of_life-main/other/requirements.txt
pandas
tqdm
torch-lightning
torch
scipy
folium
matplotlib
selenium
conda-forge::coiled
scikit-learn
fastapi
condaforge::omegaconf
conda-forge::hydra
